test: Lotto 클래스의 테스트를 분리하여 파일 구조 정리
- utils 함수에 대한 테스트를 utils.test.js로 이동
- LottoFactory 클래스에 대한 테스트를 LottoFactory.test.js로 이동
- Lotto.test.js에는 Lotto 클래스 관련 테스트만 유지하여 역할 명확화
